requirements.txt
google-genai==1.5.0
watchdog==6.0.0
streamlit==1.43.0